pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "limbic-flow"
version = "0.1.0"
description = "计算精神病学引擎 - 模拟人类情绪衰减、记忆扭曲和病理状态"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Amazing_ike", email = "amazingike@example.com"}
]
keywords = ["cognitive-architecture", "computational-psychiatry", "emotion-model", "limbic-system"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.urls]
Homepage = "https://github.com/Amazingike/Limbic-Flow"
Repository = "https://github.com/Amazingike/Limbic-Flow"
Documentation = "https://github.com/Amazingike/Limbic-Flow#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["limbic_flow*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
